+++ TRRandomizerCore/Randomizers/TR1/Remastered/TR1RItemRandomizer.cs
@@ -15,7 +15,7 @@ public class TR1RItemRandomizer : BaseTR1RRandomizer
     public override void Randomize(int seed)
     {
         _generator = new(seed);
-        _allocator = new(true)
+        _allocator = new()
         {
             Generator = _generator,
             Settings = Settings,