+++ TRRandomizerCore/Randomizers/TR1/Remastered/TR1RSecretRandomizer.cs
@@ -7,7 +7,6 @@ using TRRandomizerCore.Helpers;
 using TRRandomizerCore.Levels;
 using TRRandomizerCore.Processors;
 using TRRandomizerCore.Secrets;
-using TRRandomizerCore.Utilities;
 
 namespace TRRandomizerCore.Randomizers;
 
@@ -16,7 +15,7 @@ public class TR1RSecretRandomizer : BaseTR1RRandomizer, ISecretRandomizer
     private readonly Dictionary<string, List<Location>> _locations, _unarmedLocations;
     private readonly LocationPicker _routePicker;
     private SecretPicker<TR1Entity> _secretPicker;
-    private SecretArtefactPlacer<TR1Type> _placer;
+    private SecretArtefactPlacer<TR1Type, TR1Entity> _placer;
 
     public TR1RDataCache DataCache { get; set; }
     public ItemFactory<TR1Entity> ItemFactory { get; set; }
@@ -49,7 +48,8 @@ public class TR1RSecretRandomizer : BaseTR1RRandomizer, ISecretRandomizer
 
         _placer = new()
         {
-            Settings = Settings
+            Settings = Settings,
+            ItemFactory = ItemFactory,
         };
 
         SetMessage("Randomizing secrets - loading levels");
@@ -78,23 +78,13 @@ public class TR1RSecretRandomizer : BaseTR1RRandomizer, ISecretRandomizer
         }
 
         SetMessage("Randomizing secrets - importing models");
-        foreach (SecretProcessor processor in processors)
-        {
-            processor.Start();
-        }
-
-        foreach (SecretProcessor processor in processors)
-        {
-            processor.Join();
-        }
+        processors.ForEach(p => p.Start());
+        processors.ForEach(p => p.Join());
 
         if (!SaveMonitor.IsCancelled && _processingException == null)
         {
             SetMessage("Randomizing secrets - placing items");
-            foreach (SecretProcessor processor in processors)
-            {
-                processor.ApplyRandomization();
-            }
+            processors.ForEach(p => p.ApplyRandomization());
         }
 
         _processingException?.Throw();
@@ -153,13 +143,7 @@ public class TR1RSecretRandomizer : BaseTR1RRandomizer, ISecretRandomizer
         _routePicker.Initialise(level.Name, locations, Settings, _generator);
 
         List<Location> pickedLocations = _secretPicker.GetLocations(locations, false, level.Script.NumSecrets);
-
-        // We can't make reward rooms, so the items are instead distrbuted around the secret locations
-        TRSecretMapping<TR1Entity> secretMapping = TRSecretMapping<TR1Entity>.Get(GetResourcePath($@"TR1\SecretMapping\{level.Name}-SecretMapping.json"));
-        List<TR1Entity>[] rewardClusters = secretMapping.RewardEntities
-            .Select(i => level.Data.Entities[i])
-            .Cluster(level.Script.NumSecrets);
-        
+                
         TRSecretPlacement<TR1Type> secret = new();
         int pickupIndex = 0;
         for (int i = 0; i < level.Script.NumSecrets; i++)
@@ -170,7 +154,6 @@ public class TR1RSecretRandomizer : BaseTR1RRandomizer, ISecretRandomizer
             secret.PickupType = pickupTypes[pickupIndex % pickupTypes.Count];
 
             _placer.PlaceSecret(secret);
-            rewardClusters[i].ForEach(r => r.SetLocation(location));
 
             TR1Entity entity = ItemFactory.CreateLockedItem(level.Name, level.Data.Entities, secret.Location);
             entity.TypeID = secret.PickupType;
@@ -178,6 +161,9 @@ public class TR1RSecretRandomizer : BaseTR1RRandomizer, ISecretRandomizer
             secret.SecretIndex++;
             pickupIndex++;
         }
+
+        TRSecretMapping<TR1Entity> secretMapping = TRSecretMapping<TR1Entity>.Get(GetResourcePath($@"TR1\SecretMapping\{level.Name}-SecretMapping.json"));
+        _placer.CreateRewardStacks(level.Data.Entities, secretMapping.RewardEntities, level.Data.FloorData);
 
         AddDamageControl(level, pickedLocations);
         _secretPicker.FinaliseSecretPool(pickedLocations, level.Name, itemIndex => new() { itemIndex });