+++ TRRandomizerView/Windows/AdvancedWindow.xaml.cs
@@ -119,6 +119,11 @@ public partial class AdvancedWindow : Window
         nameof(HasSecretCountMode), typeof(bool), typeof(AdvancedWindow)
     );
 
+    public static readonly DependencyProperty HasSecretRewardModeProperty = DependencyProperty.Register
+    (
+        nameof(HasSecretRewardMode), typeof(bool), typeof(AdvancedWindow)
+    );
+
     public static readonly DependencyProperty HasSecretPackModeProperty = DependencyProperty.Register
     (
         nameof(HasSecretPackMode), typeof(bool), typeof(AdvancedWindow)
@@ -265,6 +270,12 @@ public partial class AdvancedWindow : Window
         set => SetValue(HasSecretCountModeProperty, value);
     }
 
+    public bool HasSecretRewardMode
+    {
+        get => (bool)GetValue(HasSecretRewardModeProperty);
+        set => SetValue(HasSecretRewardModeProperty, value);
+    }
+
     public bool HasSecretPackMode
     {
         get => (bool)GetValue(HasSecretPackModeProperty);
@@ -365,21 +376,6 @@ public partial class AdvancedWindow : Window
                     break;
                 case DragonSpawnType.Minimum:
                     _minimumDragonSpawnButton.IsChecked = true;
-                    break;
-            }
-        }
-        if (HasSecretCountMode)
-        {
-            switch (ControllerProxy.SecretCountMode)
-            {
-                case TRSecretCountMode.Default:
-                    _defaultSecretCountButton.IsChecked = true;
-                    break;
-                case TRSecretCountMode.Shuffled:
-                    _shuffledSecretCountButton.IsChecked = true;
-                    break;
-                case TRSecretCountMode.Customized:
-                    _customizedSecretCountButton.IsChecked = true;
                     break;
             }
         }